refactor: [hw7] project structure
