+++ src/weather/weather.module.ts
@@ -8,8 +8,8 @@ import { WeatherUrlBuilderService } from './weather-url-builder.service';
 import { WeatherApiProvider } from './weather-providers/weather-api.provider';
 import { OpenWeatherMapProvider } from './weather-providers/openweathermap.provider';
 import { CachedWeatherService } from './cached-weather.service';
-import { CacheModule } from '../cache/cache.module';
-import { MetricsService } from '../metrics/metrics.service';
+import { CacheModule } from '../infrastructure/cache/cache.module';
+import { MetricsService } from '../infrastructure/metrics/metrics.service';
 import { AbstractWeatherService, WeatherProvider } from '../interfaces/weather.interface';
 
 @Module({
